python: move EventLogEvent definition to flux.eventlog
Problem: The EventLogEvent class is defined in flux.job.event, but
the eventlog format is not specific to jobs.

Move the EventLogEvent class up a level to eventlog.py. For backwards
compatibility, retain the ability to import EventLogEvent from flux.job
and flux.job.event.
