+++ src/bindings/python/flux/job/event.py
@@ -8,9 +8,9 @@
 # SPDX-License-Identifier: LGPL-3.0
 ###############################################################
 import errno
-import json
 
 from _flux._core import ffi
+from flux.eventlog import EventLogEvent
 from flux.job._wrapper import _RAW as RAW
 from flux.kvs import WatchImplementation
 
@@ -33,45 +33,6 @@ MAIN_EVENTS = frozenset(
         "exception",
     }
 )
-
-
-class EventLogEvent(dict):
-    """
-    wrapper class for a single KVS EventLog entry
-    """
-
-    def __init__(self, event):
-        """
-        "Initialize from a string or dict eventlog event
-        """
-        if isinstance(event, str):
-            event = json.loads(event)
-        super().__init__(event)
-        if "context" not in self:
-            self["context"] = {}
-
-    def __str__(self):
-        return "{0.timestamp:<0.5f}: {0.name} {0.context}".format(self)
-
-    @property
-    def name(self):
-        return self["name"]
-
-    @property
-    def timestamp(self):
-        return self["timestamp"]
-
-    @property
-    def context(self):
-        return self["context"]
-
-    @property
-    def context_string(self):
-        if not self.context:
-            return ""
-        return json.dumps(
-            self.context, ensure_ascii=False, separators=(",", ":"), sort_keys=True
-        )
 
 
 class JobEventWatchFuture(WatchImplementation):
